[스텍, 큐, 덱] 11월 17일 - Update
