+++ 02_스텍, 큐, 덱/BOJ_10757.cpp
