+++ 02_스텍, 큐, 덱/BOJ_4949.cpp
